chore: Move components to orderhistory/
