--- core/src/main/res/drawable/ic_star.xml
@@ -1,24 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="50dp"
-    android:height="50dp"
-    android:viewportWidth="50"
-    android:viewportHeight="50">
-  <path
-      android:pathData="M25.76,4.041L30.597,13.838L41.414,15.409C42.567,15.576 43.027,16.991 42.193,17.803L34.366,25.429L36.214,36.196C36.41,37.343 35.206,38.217 34.176,37.676L24.501,32.592L14.826,37.676C13.795,38.217 12.591,37.343 12.788,36.196L14.636,25.429L6.808,17.803C5.975,16.991 6.435,15.576 7.587,15.409L18.404,13.838L23.241,4.041C23.756,2.998 25.244,2.998 25.76,4.041Z"
-      android:fillColor="#FFA227"/>
-  <path
-      android:pathData="M16.463,11.814C16.263,11.814 16.066,11.722 15.94,11.548L10.471,4.025C10.261,3.736 10.325,3.333 10.614,3.123C10.903,2.914 11.306,2.978 11.516,3.266L16.985,10.789C17.194,11.078 17.13,11.481 16.842,11.691C16.727,11.774 16.594,11.814 16.463,11.814Z"
-      android:fillColor="#FFD364"/>
-  <path
-      android:pathData="M2.646,29.968C2.373,29.968 2.12,29.794 2.032,29.521C1.922,29.182 2.107,28.818 2.446,28.708L11.295,25.834C11.635,25.724 11.999,25.91 12.109,26.249C12.219,26.588 12.033,26.952 11.694,27.062L2.845,29.936C2.779,29.957 2.712,29.968 2.646,29.968Z"
-      android:fillColor="#FFD364"/>
-  <path
-      android:pathData="M24.5,45.837C24.144,45.837 23.854,45.548 23.854,45.192V35.893C23.854,35.536 24.144,35.247 24.5,35.247C24.857,35.247 25.146,35.536 25.146,35.893V45.192C25.146,45.548 24.857,45.837 24.5,45.837Z"
-      android:fillColor="#FFD364"/>
-  <path
-      android:pathData="M46.354,29.968C46.288,29.968 46.221,29.957 46.155,29.936L37.306,27.062C36.966,26.952 36.781,26.588 36.891,26.249C37.001,25.91 37.366,25.724 37.705,25.834L46.554,28.708C46.893,28.818 47.078,29.182 46.968,29.521C46.879,29.794 46.626,29.968 46.354,29.968Z"
-      android:fillColor="#FFD364"/>
-  <path
-      android:pathData="M32.537,11.814C32.405,11.814 32.272,11.774 32.158,11.691C31.869,11.481 31.805,11.078 32.015,10.789L37.484,3.266C37.694,2.978 38.097,2.914 38.386,3.123C38.674,3.333 38.738,3.736 38.528,4.025L33.06,11.548C32.933,11.722 32.736,11.814 32.537,11.814V11.814Z"
-      android:fillColor="#FFD364"/>
-</vector>
